LPS-98399 Function to check if an editable should be highlighted moved to Utils
